Fix scheme_len and host_len checking
The old if statement to determine whether the parsed
scheme(host) length longer than the scheme(host) length
we assigned as function argument has little problem.

The old length would point to the address that colon locates
which is exactly the same as the string length of scheme(host).

e.g. if we parse "https://.."
(delim = strchr(uri, ':'));
length = (size_t)(delim - (uri + offset));
the above code will return length == 5 which is exactly the
string length of "https".

Therefore, the old one cause the wrong judgement.