consensus/tangle: decouple partial tx mask from transaction enum
In the next commit we are going to replace the field_mask_e enum
in common/mode/transaction.h with a new set of enums, to be able to
describe in a more granular way partially loaded transactions.
Thus, we need to remove the dependency of the
partial_transaction_model_e enum definition in
consensus/tangle/tangle.h from the enum to be replaced.