Common/transaction: add more granularity in mask of loaded fields
In a partially loaded transaction, each of the fields of the
iota_transaction_t structure can be partially populated, so we need
to be able to describe what is actually loaded in a more granular
way than allowed by the existing field_mask_e enum. For this
reason, the enum is being replaced by a new set of enums, each of
which indicates which sub-field in a specific field of a
transaction object is populated.
Each of the transaction_set_*() functions now sets a specific
bitmask in the corresponding enum, and each of the getter functions
verifies via assert() that the bitmask for the specific transaction
field is set.
Resolves #683.