iota_client_get_inputs enhancement
Generate one new address and fetch one address to `iota_client_get_balances`
each time. And terminate this function `iota_client_get_inputs` once it meets
the threshold (minimun needed balances)

For requests which don't assign addr_opt.total, the if statement in the loop
would examine is the current address is an unused address. If the current
address is an unused address, then break this loop.