utils: fix warnings
Declaring variable `buffer_size` and `offset` in `ssize_t`
which solves the warning that comparison between signed and
unsigned number.

Decorate `read` in `/utils/singed_files.c` with `unsigned`
in if comparison to solve the warning that comparison
between signed and unsigned number.

Set timeout in `utils/handles/socket.c`.